# go-app/swagger/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Template:
    properties:
      CreatedAt:
        format: date-time
        type: string
      UpdatedAt:
        format: date-time
        type: string
      body:
        type: string
        x-go-name: Body
      id:
        format: uint64
        type: integer
        x-go-name: ID
      name:
        type: string
        x-go-name: Name
      tags:
        items:
          type: string
        type: array
        x-go-name: Tags
      variables:
        items:
          $ref: '#/definitions/Variable'
        type: array
        x-go-name: Variables
    type: object
    x-go-package: github.com/odpf/cookies/go-app/domain
  Variable:
    properties:
      default:
        type: string
        x-go-name: Default
      description:
        type: string
        x-go-name: Description
      name:
        type: string
        x-go-name: Name
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/odpf/cookies/go-app/domain
host: localhost:3000
info:
  description: Documentation of our App API.
  title: App.
  version: 1.0.0
paths:
  /ping:
    get:
      description: Ping call
      operationId: ping
      responses:
        "200":
          $ref: '#/responses/pingResponse'
      tags:
      - ping
  /templates:
    get:
      description: 'List Templates API: This API lists all the existing templates
        with given filers in query params'
      operationId: listTemplatesRequest
      parameters:
      - description: List Template Request
        in: query
        name: tag
        type: string
        x-go-name: Tag
      responses:
        "200":
          $ref: '#/responses/listResponse'
      tags:
      - templates
    put:
      description: 'Upsert Templates API: This API helps in creating or updating a
        template with unique name'
      operationId: createTemplateRequest
      parameters:
      - description: Create template request
        in: body
        name: Body
        schema:
          $ref: '#/definitions/Template'
      responses:
        "200":
          $ref: '#/responses/templatesResponse'
      tags:
      - templates
  /templates/{name}:
    delete:
      description: 'Delete Template API: This API deletes a template given the template
        name'
      operationId: deleteTemplatesRequest
      parameters:
      - description: Delete Template Request
        in: path
        name: name
        required: true
        type: string
        x-go-name: Name
      responses:
        "200":
          $ref: '#/responses/templatesResponse'
      tags:
      - templates
    get:
      description: 'Get Template API: This API gets a template given the template
        name'
      operationId: getTemplatesRequest
      parameters:
      - description: Get Template Request
        in: path
        name: name
        required: true
        type: string
        x-go-name: Name
      responses:
        "200":
          $ref: '#/responses/templatesResponse'
      tags:
      - templates
  /templates/{name}/render:
    post:
      description: 'Render Template API: This API renders the given template with
        given values'
      operationId: renderTemplatesRequest
      parameters:
      - description: Render Template Request
        in: path
        name: name
        required: true
        type: string
        x-go-name: Name
      responses:
        "200":
          $ref: '#/responses/renderTemplatesResponse'
      tags:
      - templates
produces:
- application/json
responses:
  listResponse:
    description: List templates response
    schema:
      items:
        $ref: '#/definitions/Template'
      type: array
  pingResponse:
    description: Response body for Ping.
  renderTemplatesResponse:
    description: ""
  templatesResponse:
    description: ""
    schema:
      $ref: '#/definitions/Template'
schemes:
- http
securityDefinitions:
  basic:
    type: basic
swagger: "2.0"